requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
certifi==2020.4.5.2
chardet==3.0.4
idna==2.9
multidict==4.7.6
pkg-resources==0.0.0
requests==2.23.0
slackclient==2.7.1
urllib3==1.25.9
yarl==1.4.2
